@@ -3,13 +3,13 @@ requires = ["pdm-backend"]
 build-backend = "pdm.backend"
 
 [project]
-name = "pyfs"
+name = "pyfsviz"
 description = "Python tools for FreeSurfer visualization and QA"
 authors = [{name = "Lezlie Espana", email = "lespana@mcw.edu"}]
 license = "MIT"
 license-files = ["LICENSE"]
 readme = "README.md"
-requires-python = ">=3.9"
+requires-python = ">=3.10"
 keywords = []
 dynamic = ["version"]
 classifiers = [
@@ -18,7 +18,6 @@ classifiers = [
     "Programming Language :: Python",
     "Programming Language :: Python :: 3",
     "Programming Language :: Python :: 3 :: Only",
-    "Programming Language :: Python :: 3.9",
     "Programming Language :: Python :: 3.10",
     "Programming Language :: Python :: 3.11",
     "Programming Language :: Python :: 3.12",
@@ -41,17 +40,15 @@ dependencies = [
 ]
 
 [project.urls]
-Homepage = "https://mcw-meier-lab.github.io/pyFS"
-Documentation = "https://mcw-meier-lab.github.io/pyFS"
-Changelog = "https://mcw-meier-lab.github.io/pyFS/changelog"
-Repository = "https://github.com/mcw-meier-lab/pyFS"
-Issues = "https://github.com/mcw-meier-lab/pyFS/issues"
-Discussions = "https://github.com/mcw-meier-lab/pyFS/discussions"
-Gitter = "https://gitter.im/mcw-meier-lab/pyFS"
-Funding = "https://github.com/sponsors/l-espana"
+Homepage = "https://mcw-meier-lab.github.io/pyFSViz"
+Documentation = "https://mcw-meier-lab.github.io/pyFSViz"
+Changelog = "https://mcw-meier-lab.github.io/pyFSViz/changelog"
+Repository = "https://github.com/mcw-meier-lab/pyFSViz"
+Issues = "https://github.com/mcw-meier-lab/pyFSViz/issues"
+Discussions = "https://github.com/mcw-meier-lab/pyFSViz/discussions"
 
 [project.scripts]
-pyfs = "pyfs.cli:main"
+pyfsviz = "pyfsviz._internal.cli:main"
 
 [tool.pdm.version]
 source = "call"
