@@ -1,26 +1,23 @@
-"""pyFS package.
+"""pyFSViz package.
 
 Python tools for FreeSurfer visualization and QA
 """
 
 from __future__ import annotations
 
-from pyfs import freesurfer, reports
-from pyfs._internal.cli import get_parser, main
-from pyfs.freesurfer import (
+from pyfsviz._internal.cli import get_parser, main
+from pyfsviz.freesurfer import (
     FreeSurfer,
     get_freesurfer_colormap,
 )
-from pyfs.reports import (
+from pyfsviz.reports import (
     Template,
 )
 
 __all__: list[str] = [
     "FreeSurfer",
     "Template",
-    "freesurfer",
     "get_freesurfer_colormap",
     "get_parser",
     "main",
-    "reports",
 ]
