@@ -3,11 +3,11 @@
 # You might be tempted to import things from `__main__` later,
 # but that will cause problems: the code will get executed twice:
 #
-# - When you run `python -m pyfs` python will execute
+# - When you run `python -m pyfsviz` python will execute
 #   `__main__.py` as a script. That means there won't be any
-#   `pyfs.__main__` in `sys.modules`.
+#   `pyfsviz.__main__` in `sys.modules`.
 # - When you import `__main__` it will get executed again (as a module) because
-#   there's no `pyfs.__main__` in `sys.modules`.
+#   there's no `pyfsviz.__main__` in `sys.modules`.
 
 from __future__ import annotations
 
@@ -15,7 +15,7 @@ import argparse
 import sys
 from typing import Any
 
-from pyfs._internal import debug
+from pyfsviz._internal import debug
 
 
 class _DebugInfo(argparse.Action):
@@ -34,7 +34,7 @@ def get_parser() -> argparse.ArgumentParser:
     -------
     An argparse parser.
     """
-    parser = argparse.ArgumentParser(prog="pyfs")
+    parser = argparse.ArgumentParser(prog="pyfsviz")
     parser.add_argument("-V", "--version", action="version", version=f"%(prog)s {debug._get_version()}")
     parser.add_argument("--debug-info", action=_DebugInfo, help="Print debug information.")
     return parser
@@ -43,7 +43,7 @@ def get_parser() -> argparse.ArgumentParser:
 def main(args: list[str] | None = None) -> int:
     """Run the main program.
 
-    This function is executed when you type `pyfs` or `python -m pyfs`.
+    This function is executed when you type `pyfsviz` or `python -m pyfsviz`.
 
     Parameters
     ----------
