@@ -56,7 +56,7 @@ def _interpreter_name_version() -> tuple[str, str]:
     return "", "0.0.0"
 
 
-def _get_version(dist: str = "pyfs") -> str:
+def _get_version(dist: str = "pyfsviz") -> str:
     """Get version of the given distribution.
 
     Parameters
@@ -81,8 +81,8 @@ def _get_debug_info() -> _Environment:
         Environment information.
     """
     py_name, py_version = _interpreter_name_version()
-    packages = ["pyfs"]
-    variables = ["PYTHONPATH", *[var for var in os.environ if var.startswith("PYFS")]]
+    packages = ["pyfsviz"]
+    variables = ["PYTHONPATH", *[var for var in os.environ if var.startswith("PYFSVIZ")]]
     return _Environment(
         interpreter_name=py_name,
         interpreter_version=py_version,
